requested changes

+++ installer/R00_emba_remove.sh
@@ -29,10 +29,16 @@ R00_emba_remove() {
   pkill -f "emba.sh" || true
 
   echo -e "\\n""$ORANGE""Stopping mongod process""$NC"
-  systemctl stop mongod
+  if [[ -f "/etc/init.d/mongod" ]]; then
+    /etc/init.d/mongod stop
+  else
+    systemctl stop mongod
+  fi
+  echo -e "\\n""$ORANGE""Stopping redis-server process""$NC"
   if [[ -f "/etc/init.d/redis-server" ]]; then
-    echo -e "\\n""$ORANGE""Stopping redis-server process""$NC"
     /etc/init.d/redis-server stop
+  else
+    systemctl stop redis-server
   fi
 
   echo -e "\\n""$ORANGE""Removing redis-server packages""$NC"
